+++ docs/source/conf.py
@@ -378,6 +378,7 @@ nb_mime_priority_overrides = [
 linkcheck_ignore = [
     "https://doi.org/10.1063/1.447644",
     "https://doi.org/10.1063/1.5089550",
+    "https://doi.org/10.1063/1.5133894",
     "https://doi.org/10.1109/TQE.2020.2965810",
     "https://www.sciencedirect.com/science/article/pii/S0370157313000562",
     "dot",  # To not require graphviz in linkcheck image